feat: when the toggle is in the off state, 'no schedule' is always displayed
